+++ openfisca_uk/tools/simulation.py
@@ -1,3 +1,4 @@
+import logging
 from openfisca_uk import CountryTaxBenefitSystem
 from openfisca_uk.entities import entities
 import numpy as np
@@ -5,8 +6,10 @@ import warnings
 from openfisca_uk.entities import *
 import numpy as np
 import warnings
+from openfisca_uk.initial_setup import set_default
 from openfisca_uk.tools.parameters import backdate_parameters
-from openfisca_uk_data import DATASETS
+from openfisca_tools import ReformType
+from openfisca_uk_data import DATASETS, SynthFRS
 from openfisca_tools.microsimulation import (
     Microsimulation as GeneralMicrosimulation,
 )
@@ -32,6 +35,37 @@ class Microsimulation(GeneralMicrosimulation):
     entities = entities
     default_dataset = DEFAULT_DATASET
     post_reform = backdate_parameters()
+
+    def __init__(
+        self, reform: ReformType = (), dataset: type = None, year: int = None
+    ):
+        if dataset is None:
+            dataset = self.default_dataset
+        else:
+            dataset = dataset
+        if year is None:
+            year = self.default_year or max(dataset.years)
+        else:
+            year = year
+
+        # Check if dataset is available
+
+        if year not in dataset.years:
+            download = input(
+                f"\nYear {year} not available in dataset {dataset.name}: \n\t* Download the dataset [y]\n\t* Use the synthetic FRS (and set default) [n]\n\nChoice: "
+            )
+            if download == "y":
+                dataset.download(year)
+            else:
+                set_default(SynthFRS)
+                dataset = SynthFRS
+                if year not in dataset.years:
+                    logging.info(
+                        f"Year {year} synthetic FRS not stored, downloading..."
+                    )
+                    dataset.download(year)
+
+        super().__init__(reform=reform, dataset=dataset, year=year)
 
 
 class IndividualSim(GeneralIndividualSim):